refactor: Move code around
