+++ src/main/services/dataCollectors/GenericMerge/model/MergeScenarioResult.groovy
@@ -1,0 +1,7 @@
+package services.dataCollectors.GenericMerge.model
+
+enum MergeScenarioResult {
+    SUCCESS_WITHOUT_CONFLICTS,
+    SUCCESS_WITH_CONFLICTS,
+    TOOL_ERROR
+}
